moved Layout and TipModal to common
