+++ src/components/common/Layout.tsx
@@ -1,5 +1,5 @@
 import { ReactNode } from "react";
-import Header from "./common/Header";
+import Header from "./Header";
 
 const Layout: React.FC<{ children: ReactNode }> = ({ children }) => {
   return (
